Add an end to end test of passing arguments, storing and retrieving return value

+++ libraries/chain/webassembly/runtimes/eos-vm.cpp
@@ -148,7 +148,7 @@ class eos_vm_instantiated_module : public wasm_instantiated_module_interface {
                 iface, "env", "sync_call",
                 sync_call_ctx->sender.to_uint64_t(),
                 sync_call_ctx->receiver.to_uint64_t(),
-                static_cast<uint64_t>(sync_call_ctx->data.size())); // size_t can be uint32_t or uint64_t depending on architecture. Safely cast to uint64_t to always match the type expected by sync_call entry point.
+                static_cast<uint32_t>(sync_call_ctx->data.size()));
          };
 
          execute(context, bkend, exec_ctx, wasm_alloc, fn, true);
